training_pipeline/dev_requirements.txt
apache-airflow==2.3.4
apache-airflow-providers-docker==3.1.0
black==22.6.0
fastparquet==0.8.3
feast[redis]==0.24.0
jupyterlab==3.4.5
mlflow==1.28.0
pandas==1.4.4
pyspark==3.0.1
scikit-learn==1.1.2
